+++ RELEASES.md
@@ -4,6 +4,13 @@ Rhai Release Notes
 Version 0.19.9
 ==============
 
+Breaking changes
+----------------
+
+* `Engine::load_package` is renamed `Engine::register_global_module`.
+* `Package::get` is renamed `Package::as_shared_module`.
+* `Engine::register_module` is renamed `Engine::register_static_module`.
+
 
 Version 0.19.8
 ==============
@@ -103,7 +110,7 @@ New features
 * New `switch` statement.
 * New `do ... while` and `do ... until` statements.
 * New `Engine::gen_fn_signatures`, `Module::gen_fn_signatures` and `PackagesCollection::gen_fn_signatures` to generate a list of signatures for functions registered.
-* New `Engine::register_module` to register a module as a sub-module in the global namespace.
+* New `Engine::register_static_module` to register a module as a sub-module in the global namespace.
 * New `set_exported_global_fn!` macro to register a plugin function and expose it to the global namespace.
 * `Module::set_fn_XXX_mut` can expose a module function to the global namespace. This is convenient when registering an API for a custom type.
 * `Module::set_getter_fn`, `Module::set_setter_fn`, `Module::set_indexer_get_fn`, `Module::set_indexer_set_fn` all expose the function to the global namespace by default. This is convenient when registering an API for a custom type.
@@ -370,7 +377,7 @@ Breaking changes
 * `Engine::register_raw_fn_XXX` API shortcuts are removed.
 * `PackagesCollection::get_fn`, `PackagesCollection::contains_fn`, `Module::get_fn` and `Module::contains_fn` now take an additional `public_only` parameter indicating whether only public functions are accepted.
 * The iterator returned by `Scope::iter` now contains a clone of the `Dynamic` value (unshared).
-* `Engine::load_package` takes any type that is `Into<PackageLibrary>`.
+* `Engine::register_global_module` takes any type that is `Into<PackageLibrary>`.
 * Error in `Engine::register_custom_syntax` is no longer `Box`-ed.
 
 Housekeeping